Fixed bug where None values were being respected.
Using the add_arguments would generate a dictionary which
would result in a None value for all keys. When we went to
get those keys, it would respect that None and assume that
you wanted that value. This is not intended default behavior.
It is possible someone would want to do this in the future.
the long-term solution is probably just to add a flag to the
load_config method call to respect these none-type values.