Added format_env and format_cli flags.
These flags allow you to turn off the auto-formatting that we
do for command-line arguments and environment variables.
Currently you get all-or-nothing. Maybe in the future it might
be nice to provide someone with cli and environment formatters
so they can do whatever they want. For now, I think these are
sensible defaults based on our understanding in #7