Issue/21 (#23)
* Added item to NotFound exception.
* Moved pytest-runner from setup_requires.

This commit fixes #21 and #22 by adding the item to the 
YapconfItemNotFound error. From here, you can get all 
the information about the item, not just the name.

This was causing problems where the package would install
pytest-runner during setup time. Since the package does not
need pytest-runner to run, I removed it.