Issue/63 (#71)
* Fixed/updated sources.

* Exposed some basic properties.

* Added configuration documentation generation. #63

This commit fixes #63

It adds another call to the YapconfSpec to genrate documentation
for your configuration. It works with unicode characters. An
example of the type of documentation it can generate can be
found in `example/doc.md`

* Added documentation for #63

* Added back autodoc stuff.

This change means that the API documentation will show up
correctly again on the readthedocs page.